+++ src/suou/validators.py
@@ -22,7 +22,7 @@ _T = TypeVar('_T')
 
 def matches(regex: str | int, /, length: int = 0, *, flags=0):
     """
-    Return a function which returns true if X is shorter than length and matches the given regex.
+    Return a function which returns True if X is shorter than length and matches the given regex.
     """
     if isinstance(regex, int):
         length = regex
@@ -31,13 +31,23 @@ def matches(regex: str | int, /, length: int = 0, *, flags=0):
         return (not length or len(s) < length) and bool(re.fullmatch(regex, s, flags=flags))
     return validator
 
+
 def must_be(obj: _T | Any, typ: type[_T] | Iterable[type], message: str, *, exc = TypeError) -> _T:
     """
     Raise TypeError if the requested object is not of the desired type(s), with a nice message.
+
+    (Not properly a validator.)
     """
     if not isinstance(obj, typ):
         raise TypeError(f'{message}, not {obj.__class__.__name__!r}')
     return obj
 
 
-__all__ = ('matches', )
+def not_greater_than(y):
+    """
+    Return a function that returns True if X is not greater than (i.e. lesser than or equal to) the given value.
+    """
+    return lambda x: x <= y
+
+__all__ = ('matches', 'not_greater_than')
+