+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 + Add 7 new throwable exceptions
 + Add color utilities: `chalk` module
 + Add `.terminal` module, to ease TUI development.
++ `calendar`: add `parse_time()`.
++ Add validator `not_greater_than()`.
 
 ## 0.6.1